+++ renderer/instance/instance_v1/TerrainPass.cpp
@@ -5,7 +5,7 @@
 #include "TerrainPass.h"
 #include "LLVK_Descriptor.hpp"
 #include "VulkanRenderer.h"
-#include "renderer/shadowmap_v2/UT_ShadowMap.hpp"
+#include "renderer/public/UT_CustomRenderer.hpp"
 LLVK_NAMESPACE_BEGIN
 void TerrainGeometryContainer::buildSet() {
     const auto &device = requiredObjects.pVulkanRenderer->getMainDevice().logicalDevice;
